pyproject.toml
[project]
name = "electronics-network"
version = "0.1.0"
description = ""
authors = [
    {name = "origin",email = "yura28049119@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django = "^5.2.8"
psycopg2 = "^2.9.11"
dotenv = "^0.9.9"
djangorestframework = "^3.16.1"
django-filter = "^25.2"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"


[tool.poetry.group.lint.dependencies]
black = "^25.11.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"
isort = "^7.0.0"

